Reworked the MapAsset system; changed the placement of assets from MapRoom to the asset itself to better facilitate different placement of derived types from MapAsset. Removed the temp. asset drawing in MapManager::Draw(), and added a correct implementation of getting the MapAssets and drawing them on top of the base map.

+++ SpaceGamePrelim/Components/CenterPosComp.cpp
@@ -25,6 +25,6 @@ void CenterPosComp::Execute()
 	//	MainApplication::Instance()->GetWndHeight() / 2 - round((double)(Dim.Height() / 2))));
 
 	m_Owner->SetPosition(Vector(
-		MainApplication::Instance()->GetWndWidth() / 2,
-		MainApplication::Instance()->GetWndHeight() / 2 ));
+		(float)MainApplication::Instance()->GetWndWidth() / 2,
+		(float)MainApplication::Instance()->GetWndHeight() / 2 ));
 }
